Added bcrypt, authentication
I added the bcrypt dependency to the package.json.  New route and data
modules for authentication have been added as well as some initial
clientside javascript for the login screen. I believe that
data/customer.js module may need some additional functionality in order
to successfully allow the authentication to compare a hashed user input
to the hashed username and password residing in the database.  Currently
the only way to check the hashed user info stored in the database is to
look them up by passing a parameter, but I believe bcrypt.compare is
needed to accurately check input vs. stored data and this is only
possibly if the stored info can be retrieved first.