Update seed.js and package.json
Seed.js now creates a hashed username and password and uses this for the
John Doe customer.  The hashed version is different upon each generation
so a method will need to be created in data/customer to retrieve the
hashed info and use bcrypt.compare in the authentication module to
compare the user input to it. Simply doing a string compare with the
user input and the stored hash does not appear to be possible.