security changes and login authentication
added a number of updates for login authentication. Customer data module
has been updated so that a customer has a username and hashed password.
A customer can be retrieved by their username and then a bcrypt compare
will compare the stored hash against the hashed input to check if the
password is correct. If login is successful you are rerouted to
/customers/home which shows the login username