added new ajax POST request in login.js
Added in a new post request to login.js which adds the new customer to
the database via the signup form.  At this point, the customer will be
present in the database, and due to the POST request getting
authentication from passport, the window can be successfully redirected
to customers/home.