Customer home front end changes
Improved the response from updating financial info on customer/home;
namely, rather than a basic alert box, an alert-success or alert-danger
box will display below the submit button (remove the hidden class) after
submitting. A click anywhere on the page after this will hide the
message again. Also updated some of the placeholder text in the inputs
on signup.  "date" inputs only work on chrome and not firefox so next I
will be improving the date input to ensure the user puts the right
format of text, as well as making the "remember me" checkbox function
using local storage.